seperated the map page from app.js. still wonky tho
